pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raphael"
version = "0.1.0"
description = "Raphael AI Agent Framework"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "typer>=0.12.0",
    "gradio>=4.0.0",
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["core*", "memory*", "tools*", "agents*", "interfaces*", "config*"]

[project.scripts]
raphael = "main:app"
